+++ lib/src/date.dart
@@ -1,5 +1,6 @@
 
 import 'package:intl/intl.dart';
+import 'package:intl/date_symbols.dart';
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:swiss_knife/src/math.dart';
 
@@ -38,6 +39,15 @@ String dateFormat_YYYY_MM_dd([int time]) {
 DateTime parseDateTime(dynamic v, [DateTime def]) {
   if (v == null) return def ;
 
+  if (v is DateTime) {
+    return v ;
+  }
+
+  if (v is int) {
+    if (v == 0 && def != null) return def ;
+    return DateTime.fromMillisecondsSinceEpoch(v) ;
+  }
+
   String s ;
   if (v is String) {
     s = v ;
@@ -58,6 +68,193 @@ List<DateTime> parseDateTimeFromInlineList(dynamic s, [Pattern delimiter = ',', 
   if (s is DateTime) return [s];
   if (s is List) return s.map( (e) => parseDateTime(e) ).toList() ;
   return parseFromInlineList( s.toString() , delimiter , parseDateTime , def) ;
+}
+
+Duration parseDuration(String unit, int amount, [Duration def]) {
+  if (unit == null) return def ;
+  unit = unit.toLowerCase().trim() ;
+  if (unit.isEmpty) return def ;
+
+  amount ??= 0 ;
+
+  switch (unit) {
+    case 'year':
+    case 'years': return Duration( days: amount*365 ) ;
+
+    case 'quarter':
+    case 'quarters': return Duration( days: amount*90 ) ;
+
+    case 'month':
+    case 'months': return Duration( days: amount*30 ) ;
+
+    case 'd':
+    case 'day':
+    case 'days': return Duration( days: amount ) ;
+
+    case 'h':
+    case 'hr':
+    case 'hrs':
+    case 'hour':
+    case 'hours': return Duration( hours: amount ) ;
+
+    case 'min':
+    case 'minute':
+    case 'minutes': return Duration( minutes: amount ) ;
+
+    case 's':
+    case 'sec':
+    case 'second':
+    case 'seconds': return Duration( seconds: amount ) ;
+
+    case 'ms':
+    case 'millis':
+    case 'millisecond':
+    case 'milliseconds': return Duration( milliseconds: amount ) ;
+
+    case 'µs':
+    case 'µsec':
+    case 'us':
+    case 'usec':
+    case 'microsecond':
+    case 'microseconds': return Duration( microseconds: amount ) ;
+
+    default: return def ;
+  }
+}
+
+enum Unit {
+  Microseconds,
+  Milliseconds,
+  Seconds,
+  Minutes,
+  Hours,
+  Days,
+  Weeks,
+  Months,
+  Quarters,
+  Years,
+}
+
+Unit getUnitByIndex(int index, [Unit def]) {
+  if (index == null || index < 0 || index >= Unit.values.length ) return def ;
+  return Unit.values[index] ;
+}
+
+Unit getUnitByName(String name, [Unit def]) {
+  if (name == null) return def ;
+  name = name.toLowerCase().trim() ;
+  if (name.isEmpty) return def ;
+
+  switch (name) {
+    case 'y':
+    case 'year':
+    case 'years': return Unit.Years ;
+
+    case 'q':
+    case 'quarter':
+    case 'quarters': return Unit.Quarters ;
+
+    case 'month':
+    case 'months': return Unit.Months ;
+
+    case 'w':
+    case 'week':
+    case 'weeks': return Unit.Weeks ;
+
+    case 'd':
+    case 'day':
+    case 'days': return Unit.Days ;
+
+    case 'h':
+    case 'hr':
+    case 'hrs':
+    case 'hour':
+    case 'hours': return Unit.Hours ;
+
+    case 'm':
+    case 'min':
+    case 'minute':
+    case 'minutes': return Unit.Minutes ;
+
+    case 's':
+    case 'sec':
+    case 'second':
+    case 'seconds': return Unit.Seconds ;
+
+    case 'ms':
+    case 'milli':
+    case 'millis':
+    case 'millisecond':
+    case 'milliseconds': return Unit.Milliseconds ;
+
+    case 'µs':
+    case 'µsec':
+    case 'us':
+    case 'usec':
+    case 'micro':
+    case 'micros':
+    case 'microsecond':
+    case 'microseconds': return Unit.Microseconds ;
+
+    default: return def ;
+  }
+}
+
+Unit parseUnit(dynamic unit, [Unit def]) {
+  if (unit == null) {
+    return def ;
+  }
+  else if (unit is Unit) {
+    return unit ;
+  }
+  else if (unit is String) {
+    return getUnitByName(unit) ;
+  }
+  else if (unit is int) {
+    return getUnitByIndex(unit) ;
+  }
+  else {
+    return def;
+  }
+}
+
+int _getUnitMilliseconds(dynamic unit) {
+  if (unit == null) return null ;
+
+  var unitParsed = parseUnit(unit) ;
+
+  switch (unitParsed) {
+    case Unit.Milliseconds: return 1 ;
+    case Unit.Seconds: return 1000 ;
+    case Unit.Minutes: return 1000*60 ;
+    case Unit.Hours: return 1000*60*60 ;
+    case Unit.Days: return 1000*60*60*24 ;
+    case Unit.Weeks: return 1000*60*60*24*7 ;
+    case Unit.Months: return 1000*60*60*24*30 ;
+    case Unit.Quarters: return 1000*60*60*24*90 ;
+    case Unit.Years: return 1000*60*60*24*365 ;
+    default: return null ;
+  }
+}
+
+int getUnitAsMilliseconds(dynamic unit, [int amount = 1]) {
+  var ms = _getUnitMilliseconds(unit) ;
+  return ms * amount ;
+}
+
+double getMillisecondsAsUnit(int ms, dynamic unit, [double def]) {
+  if (ms == null) return def ;
+  if (ms == 0) return 0 ;
+
+  if (unit == null) return def ;
+
+  var unitParsed = parseUnit(unit) ;
+  if (unitParsed == null) return def ;
+
+  var unitMs = _getUnitMilliseconds(unitParsed) ;
+  var res = ms / unitMs;
+
+  return res ;
 }
 
 String getDateAmPm([int time]) {
@@ -183,6 +380,23 @@ DateTimeWeekDay getDateTimeWeekDay_from_ISO_8601_index( int weekDayIndex ) {
     case 5: return DateTimeWeekDay.Saturday ;
     case 6: return DateTimeWeekDay.Sunday ;
     default: throw ArgumentError('Invalid ISO 8601 weekDay index. Should be of range 0-6, where Monday is 0 and Sunday is 6 (Monday-to-Sunday week).');
+  }
+}
+
+DateTimeWeekDay getDateTimeWeekDayByName( String weekDayName ) {
+  if (weekDayName == null) return null ;
+  weekDayName = weekDayName.toLowerCase().trim() ;
+  if (weekDayName.isEmpty) return null ;
+
+  switch (weekDayName) {
+    case 'monday': return DateTimeWeekDay.Monday ;
+    case 'tuesday': return DateTimeWeekDay.Tuesday ;
+    case 'wednesday': return DateTimeWeekDay.Wednesday ;
+    case 'thursday': return DateTimeWeekDay.Thursday ;
+    case 'friday': return DateTimeWeekDay.Friday ;
+    case 'saturday': return DateTimeWeekDay.Saturday ;
+    case 'sunday': return DateTimeWeekDay.Sunday ;
+    default: throw ArgumentError('Invalid DateTime week day name. Should in English.');
   }
 }
 
@@ -345,5 +559,87 @@ Pair<DateTime> getDateTimeRange( DateRangeType rangeType , [DateTime now, DateTi
     default: throw UnsupportedError("Can't handle: $rangeType") ;
   }
 
+}
+
+DateTime getDateTimeStartOf(DateTime time, dynamic unit, { DateTimeWeekDay weekFirstDay , String locale } ) {
+  var unitParsed = parseUnit(unit) ;
+  if (unitParsed == null) return null ;
+
+  switch (unitParsed) {
+    case Unit.Years: return DateTime( time.year ) ;
+    case Unit.Quarters: return DateTime( time.year , (time.month~/3)*3 ) ;
+    case Unit.Weeks: {
+      weekFirstDay ??= getWeekFirstDay() ;
+      var dateTimeRange = getDateTimeRange( DateRangeType.THIS_WEEK , time, weekFirstDay ) ;
+      return dateTimeRange.a ;
+    }
+    case Unit.Months: return DateTime( time.year , time.month ) ;
+    case Unit.Days: return DateTime( time.year , time.month, time.day ) ;
+    case Unit.Hours: return DateTime( time.year , time.month, time.day, time.hour ) ;
+    case Unit.Minutes: return DateTime( time.year , time.month, time.day, time.hour, time.minute ) ;
+    case Unit.Seconds: return DateTime( time.year , time.month, time.day, time.hour, time.minute, time.second ) ;
+    default: break ;
+  }
+
+  if ('$unit'.toLowerCase().trim() == 'date') {
+    return DateTime( time.year, time.month, time.day ) ;
+  }
+
+  throw ArgumentError("Can't handle unit: $unit") ;
+}
+
+DateTime getDateTimeEndOf(DateTime time, dynamic unit, { DateTimeWeekDay weekFirstDay , String locale } ) {
+  var unitParsed = parseUnit(unit) ;
+  if (unitParsed == null) return null ;
+
+  switch (unitParsed) {
+    case Unit.Years: return DateTime( time.year , 12 , 31 , 23, 59, 59, 999 ) ;
+    case Unit.Quarters: return getDateTimeThisMonth( getDateTimeStartOf(time, unit) ).b ;
+    case Unit.Weeks: {
+      weekFirstDay ??= getWeekFirstDay() ;
+      var dateTimeRange = getDateTimeRange( DateRangeType.THIS_WEEK , time, weekFirstDay ) ;
+      return dateTimeRange.b ;
+    }
+    case Unit.Months: return getDateTimeThisMonth( DateTime( time.year , time.month , 1 ) ).b ;
+    case Unit.Days: return DateTime( time.year , time.month, time.day , 23, 59, 59, 999) ;
+    case Unit.Hours: return DateTime( time.year , time.month, time.day, time.hour , 59, 59, 999) ;
+    case Unit.Minutes: return DateTime( time.year , time.month, time.day, time.hour, time.minute , 59, 999 ) ;
+    case Unit.Seconds: return DateTime( time.year , time.month, time.day, time.hour, time.minute, time.second , 999 ) ;
+    default: break ;
+  }
+
+  if ('$unit'.toLowerCase().trim() == 'date') {
+    return getDateTimeStartOf(time, unit).add( Duration( hours: 23 , minutes: 59, seconds: 59, milliseconds: 999 )) ;
+  }
+
+  throw ArgumentError("Can't handle unit: $unit") ;
+}
+
+DateTimeWeekDay getWeekFirstDay( [String locale] ) {
+  var dateSymbols = _getLocaleDateSymbols(locale) ;
+  if (dateSymbols == null) return DateTimeWeekDay.Monday ;
+  var firstdayofweek = dateSymbols.FIRSTDAYOFWEEK;
+  var dateTimeWeekDay = getDateTimeWeekDay_from_ISO_8601_index(firstdayofweek);
+  assert(dateTimeWeekDay != null) ;
+  return dateTimeWeekDay ;
+}
+
+DateSymbols _getLocaleDateSymbols( [String locale] ) {
+  locale ??= Intl.defaultLocale ?? 'en_ISO' ;
+  var language = locale.split('_')[0];
+
+  var map = dateTimeSymbolMap() ;
+  DateSymbols dateSymbols = map[locale] ;
+  dateSymbols ??= map[ language ] ;
+
+  if (dateSymbols != null) return dateSymbols ;
+
+  for ( var entry in map.entries ) {
+    if ( entry.key.toString().startsWith(language) ) {
+      return entry.value ;
+    }
+  }
+
+  return map['en_ISO'] ;
 }